move PredefinedClasses to its own file
